Remove legacy (columns) playlist
It has been superseded by NG playlist for years. It's time to remove it.
