+++ foo_ui_columns/config_appearance.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 
-#include "playlist_view.h"
+#include "playlist_view_tfhooks.h"
 #include "config_appearance.h"
 #include "config_host.h"
 #include "tab_colours.h"
