+++ foo_ui_columns/fcl_colours.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 #include "playlist_switcher_v2.h"
-#include "playlist_view.h"
+#include "playlist_view_tfhooks.h"
 #include "config.h"
 #include "tab_colours.h"
