+++ foo_ui_columns/font_notify.cpp
@@ -7,44 +7,7 @@
 
 HFONT g_status_font = nullptr;
 HFONT g_tab_font = nullptr;
-HFONT g_header_font = nullptr;
 HFONT g_plist_font = nullptr;
-
-// {82196D79-69BC-4041-8E2A-E3B4406BB6FC}
-static const GUID font_client_cp_guid = {0x82196d79, 0x69bc, 0x4041, {0x8e, 0x2a, 0xe3, 0xb4, 0x40, 0x6b, 0xb6, 0xfc}};
-
-class font_client_cp : public cui::fonts::client {
-public:
-    const GUID& get_client_guid() const override { return font_client_cp_guid; }
-    void get_name(pfc::string_base& p_out) const override { p_out = "Legacy playlist: Items"; }
-
-    cui::fonts::font_type_t get_default_font_type() const override { return cui::fonts::font_type_items; }
-
-    void on_font_changed() const override { on_playlist_font_change(); }
-};
-
-font_client_cp::factory<font_client_cp> g_font_client_cp;
-
-void on_playlist_font_change()
-{
-    {
-        if (g_font) {
-            DeleteObject(g_font);
-            g_font = nullptr;
-        }
-
-        unsigned pcount = playlist_view::list_playlist.get_count();
-        for (unsigned m = 0; m < pcount; m++) {
-            playlist_view* p_playlist = playlist_view::list_playlist.get_item(m);
-            if (p_playlist->wnd_playlist) {
-                if (!g_font)
-                    g_font = static_api_ptr_t<cui::fonts::manager>()->get_font(font_client_cp_guid);
-                p_playlist->update_scrollbar(true);
-                RedrawWindow(p_playlist->wnd_playlist, nullptr, nullptr, RDW_INVALIDATE | RDW_UPDATENOW);
-            }
-        }
-    }
-}
 
 void font_cleanup()
 {
@@ -157,50 +120,6 @@ void on_tab_font_change()
 }
 #endif
 
-// {C0D3B76C-324D-46d3-BB3C-E81C7D3BCB85}
-static const GUID font_client_cph_guid = {0xc0d3b76c, 0x324d, 0x46d3, {0xbb, 0x3c, 0xe8, 0x1c, 0x7d, 0x3b, 0xcb, 0x85}};
-
-class font_client_cph : public cui::fonts::client {
-public:
-    const GUID& get_client_guid() const override { return font_client_cph_guid; }
-    void get_name(pfc::string_base& p_out) const override { p_out = "Legacy playlist: Column titles"; }
-
-    cui::fonts::font_type_t get_default_font_type() const override { return cui::fonts::font_type_items; }
-
-    void on_font_changed() const override { on_header_font_change(); }
-};
-
-font_client_cph::factory<font_client_cph> g_font_client_cph;
-
-void on_header_font_change()
-{
-    {
-        if (g_header_font != nullptr) {
-            unsigned count = playlist_view::list_playlist.get_count();
-            for (unsigned n = 0; n < count; n++) {
-                playlist_view* p_playlist = playlist_view::list_playlist.get_item(n);
-                if (p_playlist->wnd_header) {
-                    SendMessage(p_playlist->wnd_header, WM_SETFONT, (WPARAM)0, MAKELPARAM(0, 0));
-                }
-            }
-
-            DeleteObject(g_header_font);
-            g_header_font = nullptr;
-        }
-
-        unsigned count = playlist_view::list_playlist.get_count();
-        for (unsigned n = 0; n < count; n++) {
-            playlist_view* p_playlist = playlist_view::list_playlist.get_item(n);
-            if (p_playlist->wnd_header) {
-                if (!g_header_font)
-                    g_header_font = static_api_ptr_t<cui::fonts::manager>()->get_font(font_client_cph_guid);
-                SendMessage(p_playlist->wnd_header, WM_SETFONT, (WPARAM)g_header_font, MAKELPARAM(1, 0));
-                p_playlist->on_size();
-            }
-        }
-    }
-}
-
 int CALLBACK FontSizesProc(const LOGFONT* plf, const TEXTMETRIC* ptm, DWORD FontType, LPARAM lp);
 
 struct fontsizeinfo {
@@ -241,13 +160,11 @@ void set_font_size(bool up)
 {
     LOGFONT lf_cp, lf_ng;
     static_api_ptr_t<cui::fonts::manager> api;
-    api->get_font(font_client_cp_guid, lf_cp);
     api->get_font(pvt::g_guid_items_font, lf_ng);
 
     g_get_font_size_next_step(lf_cp, up);
     g_get_font_size_next_step(lf_ng, up);
 
-    api->set_font(font_client_cp_guid, lf_cp);
     api->set_font(pvt::g_guid_items_font, lf_ng);
 }
