+++ foo_ui_columns/font_notify.h
@@ -18,14 +18,10 @@ extern HFONT g_tab_font;
 extern HFONT g_header_font;
 extern HFONT g_plist_font;
 
-void on_playlist_font_change();
-// void on_show_sidebar_change();
 void on_show_toolbars_change();
 void on_show_status_change();
 void on_show_status_pane_change();
 void on_status_font_change();
-void on_header_font_change();
-// void on_tab_font_change();
 
 void font_cleanup();
