--- foo_ui_columns/playlist_view_dropsource.cpp
@@ -1,57 +1,0 @@
-#include "stdafx.h"
-#include "playlist_view.h"
-
-HRESULT STDMETHODCALLTYPE IDropSource_playlist::QueryInterface(REFIID iid, void** ppvObject)
-{
-    if (ppvObject == nullptr)
-        return E_INVALIDARG;
-    *ppvObject = nullptr;
-    if (iid == IID_IUnknown) {
-        AddRef();
-        *ppvObject = (IUnknown*)this;
-        return S_OK;
-    }
-    if (iid == IID_IDropSource) {
-        AddRef();
-        *ppvObject = (IDropSource*)this;
-        return S_OK;
-    }
-    return E_NOINTERFACE;
-}
-
-ULONG STDMETHODCALLTYPE IDropSource_playlist::AddRef()
-{
-    return InterlockedIncrement(&refcount);
-}
-
-ULONG STDMETHODCALLTYPE IDropSource_playlist::Release()
-{
-    LONG rv = InterlockedDecrement(&refcount);
-    if (!rv) {
-#ifdef _DEBUG
-        OutputDebugString(_T("deleting IDropSource_playlist"));
-#endif
-        delete this;
-    }
-    return rv;
-}
-
-HRESULT STDMETHODCALLTYPE IDropSource_playlist::QueryContinueDrag(BOOL fEscapePressed, DWORD grfKeyState)
-{
-    if (fEscapePressed || (p_playlist->m_rmb_is_dragging && grfKeyState & MK_LBUTTON)
-        || (!p_playlist->m_rmb_is_dragging && ((grfKeyState & MK_RBUTTON) || !(grfKeyState & MK_CONTROL)))) {
-        return DRAGDROP_S_CANCEL;
-    }
-    if ((p_playlist->m_rmb_is_dragging && !(grfKeyState & MK_RBUTTON))
-        || (!p_playlist->m_rmb_is_dragging && !(grfKeyState & MK_LBUTTON))) {
-        return DRAGDROP_S_DROP;
-    }
-    return S_OK;
-}
-
-HRESULT STDMETHODCALLTYPE IDropSource_playlist::GiveFeedback(DWORD dwEffect)
-{
-    return DRAGDROP_S_USEDEFAULTCURSORS;
-}
-
-IDropSource_playlist::IDropSource_playlist(playlist_view* playlist) : refcount(0), p_playlist(playlist) {}
