--- foo_ui_columns/sort.h
@@ -1,10 +1,0 @@
-#ifndef _COLUMNS_SORT_H_
-#define _COLUMNS_SORT_H_
-
-#include "common.h"
-
-void create_image_list();
-
-#include "extern.h"
-
-#endif
